requirements.txt
gradio
langrs
torch
rasterio
geopandas
shapely
Pillow
